// src/types.ts
export interface Course {
  id: string;
  code: string;
  title: string;
  description: string;
  duration: string;
  fees: number;
  thumbnail: string;
  faculty: string; // e.g., "School of Computing", "Engineering"
  active: boolean;
}
// ...
export interface Grade {
  cat: number;   // Max 30
  exam: number;  // Max 70
}

export interface Student {
  id: string;
  name: string;
  email: string;
  phone: string;
  admissionNo: string;
  cohort: string; // e.g., "2024 Intake"
  enrolledUnits: string[]; // Code list
  grades: Record<string, Grade>; // subjectCode -> Grade
  ledger: Invoice[];
  payments: Payment[];
  attendance?: Record<string, number>; // subjectCode -> Attendance percentage (e.g. 0-100)
  avatar?: string;
  passcode?: string;
}

export interface Invoice {
  id: string;
  invoiceNo: string;
  description: string;
  amount: number;
  date: string;
  status: 'unpaid' | 'paid';
}

export interface Payment {
  id: string;
  amount: number;
  invoiceId: string;
  studentId: string;
  paymentMethod: 'M-Pesa' | 'Bank Transfer' | 'Card';
  transactionId: string;
  date: string;
  status: 'unreconciled' | 'reconciled';
}